refactor: relocate i18nextScannerService to a new directory
Move i18nextScannerService.ts to i18nextScannerService directory for 
better module organization. This enhances maintainability by aligning 
with our file structure conventions and improves code navigation.
