+++ src/lib/services/i18nextScannerService/i18nextScannerService.ts
@@ -3,10 +3,10 @@ import sort from 'gulp-sort';
 import I18nextScanner from 'i18next-scanner';
 import vfs from 'vinyl-fs';
 
-import I18nextScannerModuleConfiguration from '../entities/configuration/modules/i18nextScanner/i18nextScannerModuleConfiguration';
-import ConfigurationStoreManager from '../stores/configuration/configurationStoreManager';
-import { I18nextScannerOptions } from '../types/i18nextScannerOptions';
-import { getProjectRootFolder } from '../utilities/filePathUtilities';
+import I18nextScannerModuleConfiguration from '../../entities/configuration/modules/i18nextScanner/i18nextScannerModuleConfiguration';
+import ConfigurationStoreManager from '../../stores/configuration/configurationStoreManager';
+import { I18nextScannerOptions } from '../../types/i18nextScannerOptions';
+import { getProjectRootFolder } from '../../utilities/filePathUtilities';
 
 /**
  * Service for scanning code using i18next-scanner.
